[tests] Fixed the cases where tests_start_mpfr was called too late, i.e. after potential output (to stdout).   * tremquo.c: moved the call to tests_start_mpfr to the beginning.   * tset_float128.c, tset_ld.c: removed the optional, obsolete code     setting the FPU control word; MPFR_FPU_PREC can be used instead     (see tests.c).
[tests/tests.c] Make stderr unbuffered again.
Note for the history: both stdout and stderr were made unbuffered in
r2361, then all stderr were changed to stdout (and the corresponding
setbuf on stderr removed) in r2449 for consistency, but stderr was
used later for specific cases (e.g. memory allocation errors or when
stdout is already used for something else), and it is also used for
assertion failures. Thus the setbuf on stderr is still needed.

(merged changesets r12185-12186 from the trunk)

git-svn-id: svn://scm.gforge.inria.fr/svnroot/mpfr/branches/4.0@12187 280ebfd0-de03-0410-8827-d642c229c3f4